Add files via upload
Restaurant-Billing-Sysytem
This is a terminal based prototype of a restaurant billing system.

Introduction
This program be used in restaurants to calculate the bill of every table(For display purpose , number of tables = 10). Every table is an object of class,stored in an array of objects. It is made using OOP Principles.

Concepts used:-
Classes & Objects
Static Methods
Array of Objects
String Formatting & more